Complete matrix utils
The CL adds all the methods existent in matrix_utils.dart.
It also changes Matrix and Vector classes to use Doubles instead of
Floats everywhere (dart seems to rely on Doubles, Romain's
implementation was using Floats).
It also fixes the inc() and dec() operators to comply with kotlin's
practice that the operation should be immutable.
Tests will come in a further CL.

Test: not yet

+++ ui/port/src/main/java/androidx/ui/rendering/view/ViewConfiguration.kt
@@ -13,7 +13,7 @@ data class ViewConfiguration(
 
     // / Creates a transformation matrix that applies the [devicePixelRatio].
     fun toMatrix(): Matrix4 {
-        return Matrix4.diagonal3Values(devicePixelRatio.toFloat(), devicePixelRatio.toFloat(), 1f)
+        return Matrix4.diagonal3Values(devicePixelRatio, devicePixelRatio, 1.0)
     }
 
     override fun toString(): String {
